--- domain/src/main/java/project/side/ikdaman/domain/repository/BookApiRepository.kt
@@ -1,11 +1,0 @@
-package project.side.ikdaman.domain.repository
-
-import kotlinx.coroutines.flow.Flow
-import project.side.ikdaman.domain.model.AddBookItem
-import project.side.ikdaman.domain.model.ApiResult
-import project.side.ikdaman.domain.model.HomeBookItem
-
-interface BookApiRepository {
-    fun getBooks(): Flow<ApiResult<List<HomeBookItem>>>
-    fun postBook(addBookItem: AddBookItem): Flow<ApiResult<Unit>>
-}
