+++ domain/src/main/java/project/side/ikdaman/domain/usecase/PostBookUseCase.kt
@@ -1,9 +1,9 @@
 package project.side.ikdaman.domain.usecase
 
 import project.side.ikdaman.domain.model.AddBookItem
-import project.side.ikdaman.domain.repository.BookApiRepository
+import project.side.ikdaman.domain.repository.MyBooksApiRepository
 import javax.inject.Inject
 
-class PostBookUseCase @Inject constructor(private val bookApiRepository: BookApiRepository) {
-    operator fun invoke(addBookItem: AddBookItem) = bookApiRepository.postBook(addBookItem)
+class PostBookUseCase @Inject constructor(private val myBooksApiRepository: MyBooksApiRepository) {
+    operator fun invoke(addBookItem: AddBookItem) = myBooksApiRepository.postBook(addBookItem)
 }
